refactor(commands): decompose msgHandler god function; help becomes a real command
msgHandler was a ~188-line god function: parse guard, group gating, admin
resolution, active-command filtering (duplicated with initCommands), an entire
~80-line inlined help feature that bypassed the normal command path, lookup,
cooldown, in-flight dedup, exec, and PostgreSQL logging soldered into a finally.

Decompose:
- src/commands/registry.ts now owns allCommands + resolveActiveCommands; the
  ACTIVE_COMMANDS filter lived in two places (initCommands + msgHandler) and is
  now one function used by both.
- help is a normal IRegister (src/commands/help/register.ts) instead of a
  special case inside the dispatcher. It flows through the same lookup / cooldown
  / exec / logging path as every other command, so the 80-line help branch and
  its special-casing are deleted from msgHandler. (help declares no timesInterval
  → no cooldown, same as before; it uses a dynamic import of the registry to keep
  the registry<->help dependency one-directional, matching welcome.ts.)
- PG command logging extracted from the finally into logCommandExecution().

msgHandler is now a thin dispatcher: guard → resolve → find → admin gate →
cooldown → dedup → exec → log. No behavior change beyond help now being logged
like any other command.

Add registry tests (guards the import cycle resolves with no undefined entries;
verifies help registration and ACTIVE_COMMANDS filtering).

Verified: tsc clean, build clean (no circular-dependency warning), 258 tests
passing.
