+++ src/commands/index.ts
@@ -5,87 +5,15 @@ import {
     MessageEvent,
     ParamsType,
 } from '../types';
-import {
-    AnalyticsCommandRegister,
-    MapsCommandRegister,
-    ServersCommandRegister,
-    WhereIsCommandRegister,
-    PlayersCommandRegister,
-    ServerOverviewCommandRegister,
-} from './servers/register';
-import { RollCommandRegister } from './roll/register';
 import { logger } from '../utils/logger';
 import { RemoteService } from '../services/remote.service';
-import {
-    getCommandParams,
-    getFirstCommand,
-    parseIgnoreSpace,
-} from '../utils/cmd';
-import { checkTimeIntervalValid, getStaticHttpPath } from '../utils/cmdreq';
-import { FuckCommandRegister } from './fuck/register';
-import { SetuCommandRegister } from './setu/register';
-import { TouhouCommandRegister } from './touhou/register';
-import { WaifuCommandRegister } from './waifu/registers';
-import { OnePtCommandRegister } from './1pt/register';
-import { NekoCommandRegister } from './neko/register';
-import { WebsiteCommandRegister } from './website/register';
-import {
-    TDollCommandRegister,
-    TDollSkinCommandRegister,
-} from './tdoll/register';
-import {
-    QACommandRegister,
-    QADefineCommandRegister,
-    QADeleteCommandRegister,
-} from './qa/register';
-import { VersionCommandRegister } from './version/register';
-import {
-    LogCommandRegister,
-    LogSelfRegister,
-    Log7CommandRegister,
-} from './log/register';
-import { AiCommandRegister } from './ai/register';
+import { getCommandParams, getFirstCommand } from '../utils/cmd';
+import { checkTimeIntervalValid } from '../utils/cmdreq';
 import { PostgreSQLService, CmdData } from '../services/postgresql.service';
-import { CanvasImgService } from '../services/canvasImg.service';
-import { HelpCanvas, type HelpCanvasModel } from './help/canvas/helpCanvas';
-import { CheckCommandRegister } from './check/register';
-
-const allCommands: IRegister[] = [
-    FuckCommandRegister,
-    ServersCommandRegister,
-    WhereIsCommandRegister,
-    AnalyticsCommandRegister,
-    ServerOverviewCommandRegister,
-    MapsCommandRegister,
-    PlayersCommandRegister,
-    RollCommandRegister,
-    SetuCommandRegister,
-    TouhouCommandRegister,
-    WaifuCommandRegister,
-    OnePtCommandRegister,
-    NekoCommandRegister,
-    WebsiteCommandRegister,
-    QACommandRegister,
-    QADefineCommandRegister,
-    QADeleteCommandRegister,
-    VersionCommandRegister,
-    LogCommandRegister,
-    LogSelfRegister,
-    Log7CommandRegister,
-    AiCommandRegister,
-    TDollCommandRegister,
-    TDollSkinCommandRegister,
-    CheckCommandRegister,
-];
+import { resolveActiveCommands } from './registry';
 
 export const initCommands = async (env: GlobalEnv) => {
-    const activeCommands = env.ACTIVE_COMMANDS
-        ? new Set(env.ACTIVE_COMMANDS)
-        : null;
-
-    const filteredCommands = allCommands.filter(
-        (cmd) => !activeCommands || activeCommands.has(cmd.name),
-    );
+    const filteredCommands = resolveActiveCommands(env);
 
     await Promise.all(
         filteredCommands.map(async (cmd) => {
@@ -123,6 +51,41 @@ const formatCooldownSeconds = (remainingMs?: number) => {
     return Math.ceil(remainingMs / 1000);
 };
 
+/**
+ * 记录命令执行数据到 PostgreSQL(可选; 未配置或失败都不影响主流程)。
+ */
+const logCommandExecution = async (
+    env: GlobalEnv,
+    hitCommand: IRegister,
+    params: ParamsType,
+    event: MessageEvent,
+    receivedTime: Date,
+    responseTime: Date | undefined,
+): Promise<void> => {
+    if (!(env.PG_HOST && env.PG_DB && env.PG_USER)) {
+        return;
+    }
+
+    try {
+        const elapseTime = responseTime
+            ? responseTime.getTime() - receivedTime.getTime()
+            : 0;
+        const cmdData: CmdData = {
+            cmd: hitCommand.name,
+            params: Array.from(params.keys()).join(' '),
+            user_id: event.user_id,
+            group_id: event.group_id || 0,
+            received_time: receivedTime,
+            response_time: responseTime,
+            elapse_time: elapseTime,
+        };
+        await PostgreSQLService.getInst().insertCmdData(cmdData);
+    } catch (dbError) {
+        // 数据记录失败不影响主流程，仅记录日志
+        logger.error('[cmd] Failed to insert cmd data', dbError);
+    }
+};
+
 export const msgHandler = async (env: GlobalEnv, event: MessageEvent) => {
     const msgRaw = event.message;
     if (typeof msgRaw !== 'string') {
@@ -142,97 +105,8 @@ export const msgHandler = async (env: GlobalEnv, event: MessageEvent) => {
     logger.info('> Got bot msg', msg);
 
     const firstCommand = getFirstCommand(msg);
-
     const isAdminUser = env.ADMIN_QQ_LIST.some((qq) => event.user_id === qq);
-
-    const activeCommands = env.ACTIVE_COMMANDS
-        ? new Set(env.ACTIVE_COMMANDS)
-        : null;
-    const avaliableCommands = allCommands.filter(
-        (c) => !activeCommands || activeCommands.has(c.name),
-    );
-
-    // help:
-    if (firstCommand === 'help' || firstCommand === 'h') {
-        const params = parseIgnoreSpace(['#help', '#h'], msg);
-        const query = params.keys().next().value;
-
-        const prefix = env.START_MATCH || '#';
-        const visibleCommands = avaliableCommands.filter(
-            (c) => !c.isAdmin || isAdminUser,
-        );
-
-        const outputFile = `help_${event.group_id ? event.group_id : 'private'}_${event.user_id}.png`;
-
-        let model: HelpCanvasModel;
-        if (query) {
-            const hitCommand = visibleCommands.find(
-                (c) => c.name === query || c.alias === query,
-            );
-            if (hitCommand) {
-                model = {
-                    mode: 'detail',
-                    prefix,
-                    name: hitCommand.name,
-                    alias: hitCommand.alias,
-                    description: hitCommand.description,
-                    hints: hitCommand.hint ?? [],
-                };
-            } else {
-                model = { mode: 'not_found', prefix, query };
-            }
-        } else {
-            model = {
-                mode: 'list',
-                prefix,
-                items: visibleCommands.map((c) => ({
-                    name: c.name,
-                    alias: c.alias,
-                    description: c.description,
-                })),
-            };
-        }
-
-        let fallbackText = '';
-        try {
-            if (env.OUTPUT_BG_IMG) {
-                await CanvasImgService.getInstance().addImg(
-                    env.OUTPUT_BG_IMG,
-                    true,
-                );
-            }
-            await new HelpCanvas(model, outputFile).render();
-            const cqOutput = `[CQ:image,file=${getStaticHttpPath(
-                env,
-                outputFile,
-            )},cache=0,c=8]`;
-            await quickReply(event, cqOutput);
-        } catch (err) {
-            logger.error('[help] render failed', err);
-
-            if (query) {
-                const hitCommand = visibleCommands.find(
-                    (c) => c.name === query || c.alias === query,
-                );
-                if (hitCommand) {
-                    fallbackText = `${prefix}${hitCommand.name}${hitCommand.alias ? `(${hitCommand.alias})` : ''}: 帮助列表\n\n`;
-                    hitCommand.hint?.forEach((h) => {
-                        fallbackText += `${h}\n\n`;
-                    });
-                } else {
-                    fallbackText = '未找到对应命令\n';
-                }
-            } else {
-                fallbackText = '帮助列表: \n';
-                visibleCommands.forEach((c) => {
-                    fallbackText += `${prefix}${c.name}${c.alias ? `(${c.alias})` : ''}: ${c.description}\n\n`;
-                });
-            }
-
-            await quickReply(event, fallbackText);
-        }
-        return;
-    }
+    const avaliableCommands = resolveActiveCommands(env);
 
     const hitCommand = avaliableCommands.find(
         (c) => c.name === firstCommand || c.alias === firstCommand,
@@ -245,69 +119,51 @@ export const msgHandler = async (env: GlobalEnv, event: MessageEvent) => {
         return;
     }
 
-    if (firstCommand === hitCommand.name || firstCommand === hitCommand.alias) {
-        const params: ParamsType = hitCommand.parseParams
-            ? hitCommand.parseParams(msg)
-            : getCommandParams(msg, hitCommand.defaultParams);
-        const ctx: MsgExecCtx = {
-            msg,
+    const params: ParamsType = hitCommand.parseParams
+        ? hitCommand.parseParams(msg)
+        : getCommandParams(msg, hitCommand.defaultParams);
+    const ctx: MsgExecCtx = {
+        msg,
+        env,
+        event,
+        params,
+        reply: async (replyMsg: string) => {
+            await quickReply(event, replyMsg);
+        },
+    };
+
+    const cdRes = checkTimeIntervalValid(hitCommand, event);
+    if (!cdRes.success) {
+        await ctx.reply(
+            `命令冷却中，请 ${formatCooldownSeconds(cdRes.remainingMs)} 秒后再试`,
+        );
+        return;
+    }
+
+    if (handlingRequestSet.has(event.user_id)) {
+        await ctx.reply('上一条命令仍在处理中，请稍后再试');
+        return;
+    }
+    handlingRequestSet.add(event.user_id);
+
+    const receivedTime = new Date();
+    let responseTime: Date | undefined;
+
+    try {
+        await hitCommand.exec(ctx);
+        responseTime = new Date();
+    } catch (e) {
+        logger.error(e);
+        responseTime = new Date();
+    } finally {
+        handlingRequestSet.delete(event.user_id);
+        await logCommandExecution(
             env,
-            event,
+            hitCommand,
             params,
-            reply: async (msg: string) => {
-                await quickReply(event, msg);
-            },
-        };
-
-        const cdRes = checkTimeIntervalValid(hitCommand, event);
-        if (!cdRes.success) {
-            await ctx.reply(
-                `命令冷却中，请 ${formatCooldownSeconds(cdRes.remainingMs)} 秒后再试`,
-            );
-            return;
-        }
-
-        if (handlingRequestSet.has(event.user_id)) {
-            await ctx.reply('上一条命令仍在处理中，请稍后再试');
-            return;
-        }
-        handlingRequestSet.add(event.user_id);
-
-        const receivedTime = new Date();
-        let responseTime: Date | undefined;
-
-        try {
-            await hitCommand.exec(ctx);
-            responseTime = new Date();
-        } catch (e) {
-            logger.error(e);
-            responseTime = new Date();
-        } finally {
-            handlingRequestSet.delete(event.user_id);
-
-            if (!(env.PG_HOST && env.PG_DB && env.PG_USER)) {
-                return;
-            }
-
-            // 记录命令执行数据（失败不影响主流程）
-            try {
-                const elapseTime = responseTime
-                    ? responseTime.getTime() - receivedTime.getTime()
-                    : 0;
-                const cmdData: CmdData = {
-                    cmd: hitCommand.name,
-                    params: Array.from(params.keys()).join(' '),
-                    user_id: event.user_id,
-                    group_id: event.group_id || 0,
-                    received_time: receivedTime,
-                    response_time: responseTime,
-                    elapse_time: elapseTime,
-                };
-                await PostgreSQLService.getInst().insertCmdData(cmdData);
-            } catch (dbError) {
-                // 数据记录失败不影响主流程，仅记录日志
-                logger.error('[cmd] Failed to insert cmd data', dbError);
-            }
-        }
+            event,
+            receivedTime,
+            responseTime,
+        );
     }
 };
