+++ src/commands/registry.test.ts
@@ -1,0 +1,30 @@
+import { describe, it, expect } from 'vitest';
+import { allCommands, resolveActiveCommands } from './registry';
+import { GlobalEnv } from '../types';
+
+describe('command registry', () => {
+    it('has no undefined entries (guards the registry<->help import cycle)', () => {
+        expect(allCommands.length).toBeGreaterThan(0);
+        expect(allCommands.every((c) => c && typeof c.name === 'string')).toBe(
+            true,
+        );
+    });
+
+    it('registers help as a normal command (name + alias)', () => {
+        const help = allCommands.find((c) => c.name === 'help');
+        expect(help).toBeDefined();
+        expect(help?.alias).toBe('h');
+        expect(typeof help?.exec).toBe('function');
+    });
+
+    it('resolveActiveCommands returns all when ACTIVE_COMMANDS is empty', () => {
+        const env = {} as GlobalEnv;
+        expect(resolveActiveCommands(env)).toHaveLength(allCommands.length);
+    });
+
+    it('resolveActiveCommands filters by ACTIVE_COMMANDS', () => {
+        const env = { ACTIVE_COMMANDS: ['help', 'servers'] } as GlobalEnv;
+        const names = resolveActiveCommands(env).map((c) => c.name);
+        expect(names.sort()).toEqual(['help', 'servers']);
+    });
+});
